# axon/swagger/project-api.json
{
    "info": {
        "description": "Project API documents are on [document page](/).",
        "title": "Project API",
        "version": "20230701"
    },
    "servers": [
        {
            "url": "/",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/security/verification": {
            "post": {
                "tags": [
                    "Platform"
                ],
                "security": [
                    {
                        "BearerAuthentication": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "openapi": "3.1.0",
    "components": {
        "securitySchemes": {
            "BearerAuthentication": {
                "type": "http",
                "schema": "bearer"
            }
        }
    }
}